requirements.txt
streamlit>=1.41.0
pandas
plotly
python-dotenv>=1.0.0
pydantic>=2.0.0
bcrypt>=4.0.1
cryptography>=41.0.0
openai>=1.0.0
googlesearch-python>=1.2.3
gnews>=0.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
openpyxl>=3.1.2
reportlab>=4.0.8
wordcloud>=1.9.2
matplotlib>=3.8.2
seaborn>=0.13.1 